cli: add loglevel [get|set] commands to the heketi-cli tool
Add new command line commands that allows an administrator
working with heketi to view/change the internal logging of the
server.
Running `heketi-cli loglevel get` will display the current
log level.
Running `heketi-cli loglevel set debug` will enable debug logging
in the server.

NOTE: The api is designed to support examining and setting the log
level of multiple loggers, but only really supports the glusterfs
logger currently. The CLI hides this as an implemetation detail and
only sets/displays the glusterfs logger. In the future we could
add a --logger switch or something to allow this to be changed.

Signed-off-by: John Mulligan <jmulligan@redhat.com>