apps: prevent wasteful creation of a block-hosting volume of default size
If a block volume is requested of size larger than the configured
BlockHostingVolumeSize, and a large enough block hosting volume
has not been created manually, then heketi would still create a
block hosting volume of default size and subsequently fail the
block-volume creation in the executor.

This fixes it by checking the sizes and denying the block-volume
create request with a meaningful error message.

Closes #1089

Signed-off-by: Michael Adam <obnox@redhat.com>