executor: don't consider expand as failed if rebalance fails
A volume expand operation consists of
1. creating new bricks
2. attaching new bricks to volume using gluster cli
3. performing rebalance operation to make new and existing files to use
newly added bricks.

The possibility of a rebalance command failing is high as it fails if even one
brick belonging to the volume is down. Hence, we choose to consider
VolumeExpand operation to be successful if steps 1 and 2 succeed. A log
message is printed making admin aware of manual action required to
complete volume expansion.

Co-authored-by: Madhu Rajanna <mrajanna@redhat.com>
Signed-off-by: Raghavendra Talur <rtalur@redhat.com>